Added a proxy serial console fixture
This fixture uses real SerialConsole object, but hooked up to a pseudo tty
This allows the SerialConsole to behave as if it is talking to a real serial
port. The input and output for the console is actually sent to a file descriptor
the system. This can be written to and read from to emulate interactions with a
real serial port.