Added pytest fixture to allow use of temporary files in tests
This should allow for the Pluma CLI to be tested using config file contents
declared inline as strings, rather than having to have an actual file exist
for each test (scales badly).

The intention is for all new tests that need to use config files use this
fixture instead of real filesystem files.