Add failing test to expose bug in OsFile test utility class
A bug was found where if PexpectEngine() was opened with a pseudo terminal
file descriptior (from conftest.pty_pair & OsFile), data sent to the primary
pty is not readable on the sencond pty unless a newline is also sent.

The test exposes this bug by using the engine.send() method rather than
engine.send_line().
I've done additional testing outside of this in the python repl and found
that it is indeed the case that the following does not work:

```python
>>> import os
>>> import pty

>>> fst, snd = pty.openpty()
>>> os.write(fst, b'Hello\n')
6

\# This bit works, note the newline on the string written
>>> os.read(snd, 6)                                                                                  │
b'Hello\n'

>>> os.write(fst, b'Hello')
5

>>> os.read(snd, 5)
\# Hangs forever waiting for data to read
```